🤓 move /lib to /api/lib
MVP
